requirements.txt

# --------------------
# Backend (FastAPI)
# --------------------
fastapi==0.115.2
uvicorn==0.30.6
pandas==2.2.3
requests==2.32.3
python-multipart==0.0.9
python-dotenv
groq
sentence-transformers
joblib
matplotlib
scikit-learn

# --------------------
# Frontend (Streamlit)
# --------------------
streamlit==1.39.0
plotly==5.24.1

